arm: make the tag storage location configurable and describe it to the guest
A proposed extension to the Linux kernel requires the kernel to know how
MTE allocation tags are stored, as well as their storage location if they
are exposed in physical memory. The proposal is to add the arm,mte-alloc
property to all memory nodes whose tag storage location is either not
accessible via regular loads and stores to physical memory, or is fully
described by /reserved-memory/ child nodes with the compatible property
being equal to arm,mte-tag-storage.

At present, QEMU stores the tags in a location directly inaccessible to
the guest, so this patch simply adds the arm,mte-alloc property to the
memory nodes by default.

This patch also allows emulating an implementation that uses physical
memory for allocation tag storage by passing mte-shared-alloc=on as part
of the -machine virt argument. This causes the tag storage location to
be described via a /reserved-memory/ child node.

This patch is being shared for discussion purposes only, and should not
land upstream until the device tree extensions have been finalized.

Signed-off-by: Peter Collingbourne <pcc@google.com>